changed all occurances of "long long" into int64_t
    - makes sure integers have save width across platforms
    - addresses issue eudaq/eudaq#61

+++ gui/include/euRun.hh
@@ -146,6 +146,6 @@ signals:
     status_t m_status;
     int m_prevtrigs;
     double m_prevtime, m_runstarttime;
-    long long m_filebytes;
+    int64_t m_filebytes;
     bool dostatus;
 };
